aggiunta schermata punteggio e timer nel gioco, risolti bug scorescreen, implementata scrittura del tempo nel file

+++ TetrisGame/ScoreScreen.hpp
@@ -19,7 +19,7 @@ class ScoreScreen {
 protected:
     WINDOW  *score_screen;
     p_giocata scores;
-    int scroll_index;
+    float scroll_index;
     bool open;
     int line;
     int list_length;
